pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "polymarket-insider-detector"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27.0",
    "websockets>=12.0",
    "scipy>=1.12.0",
    "numpy>=1.26.0",
    "networkx>=3.2",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dashboard = ["streamlit>=1.31.0", "plotly>=5.18.0"]
dev = ["pytest>=8.0", "ruff>=0.3.0"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
